+++ nems/gui/db_browser.py
@@ -36,6 +36,20 @@ import nems
 configfile = os.path.join(nems.get_setting('SAVED_SETTINGS_PATH') + '/gui.ini')
 nems_root = os.path.abspath(nems.get_setting('SAVED_SETTINGS_PATH') + '/../../')
 
+# TEMP ERROR CATCHER
+# Back up the reference to the exceptionhook
+sys._excepthook = sys.excepthook
+
+def my_exception_hook(exctype, value, traceback):
+    # Print the error and traceback
+    print(exctype, value, traceback)
+    # Call the normal Exception hook after
+    sys._excepthook(exctype, value, traceback)
+    sys.exit(1)
+
+# Set the exception hook to our wrapping function
+sys.excepthook = my_exception_hook
+
 def load_settings(m):
 
     config = ConfigParser()
